requirements.txt
# ── Django Core ───────────────────────────────────────────
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.6.0
python-decouple==3.8

# ── Database ──────────────────────────────────────────────
mysqlclient==2.2.0

# ── Deep Learning ─────────────────────────────────────────
tensorflow-intel==2.15.0
keras==2.15.0
ml-dtypes==0.2.0

# ── Computer Vision ───────────────────────────────────────
opencv-python==4.8.1.78
opencv-contrib-python==4.11.0.86
mediapipe==0.10.9

# ── Machine Learning ──────────────────────────────────────
scikit-learn==1.3.2
joblib==1.3.2

# ── Data Processing ───────────────────────────────────────
numpy==1.26.4
pandas==2.1.4
scipy==1.11.4

# ── Visualization ─────────────────────────────────────────
matplotlib==3.8.2
Pillow==10.1.0

# ── TensorFlow Dependencies ───────────────────────────────
absl-py==2.1.0
flatbuffers==23.5.26
protobuf==3.20.3
grpcio==1.59.3
h5py==3.10.0
opt-einsum==3.3.0
gast==0.5.4
google-pasta==0.2.0
termcolor==2.4.0
wrapt==1.14.1
astunparse==1.6.3
typing-extensions==4.8.0
packaging==23.2

# ── Channels (WebSocket) ──────────────────────────────────
channels==4.0.0

# ── Utilities ─────────────────────────────────────────────
six==1.16.0
python-dateutil==2.8.2
pytz==2023.3
sqlparse==0.4.4
asgiref==3.7.2
